+++ packages/myst-cli/src/utils/git.ts
@@ -10,6 +10,15 @@ export async function checkFolderIsGit(): Promise<boolean> {
   }
 }
 
+export async function checkIgnore(pattern: string): Promise<boolean> {
+  try {
+    await makeExecutable(`git check-ignore ${pattern}`, null)();
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
 export async function checkAtGitRoot(): Promise<boolean> {
   return await fsExists('.git');
 }