+++ packages/myst-directives/src/dropdown.ts
@@ -1,4 +1,5 @@
 import type { DirectiveSpec, DirectiveData, GenericNode } from 'myst-common';
+import { addCommonDirectiveOptions, labelDirectiveOption } from './utils.js';
 
 export const dropdownDirective: DirectiveSpec = {
   name: 'dropdown',
@@ -6,8 +7,11 @@ export const dropdownDirective: DirectiveSpec = {
     type: 'myst',
   },
   options: {
+    ...labelDirectiveOption('dropdown'),
+    // TODO: Add enumeration in future
     open: {
       type: Boolean,
+      doc: 'When true, the dropdown starts open.',
     },
     // Legacy options we may want to implement:
     // color
@@ -29,12 +33,12 @@ export const dropdownDirective: DirectiveSpec = {
       children.push({ type: 'summary', children: data.arg as GenericNode[] });
     }
     children.push(...(data.body as GenericNode[]));
-    return [
-      {
-        type: 'details',
-        open: data.options?.open,
-        children,
-      },
-    ];
+    const details = {
+      type: 'details',
+      open: data.options?.open,
+      children,
+    };
+    addCommonDirectiveOptions(data, details);
+    return [details];
   },
 };