+++ packages/myst-directives/src/math.ts
@@ -1,26 +1,17 @@
 import type { DirectiveSpec, DirectiveData, GenericNode } from 'myst-common';
-import { normalizeLabel } from 'myst-common';
+import { addCommonDirectiveOptions, commonDirectiveOptions } from './utils.js';
 
 export const mathDirective: DirectiveSpec = {
   name: 'math',
   options: {
-    label: {
-      type: String,
-      alias: ['name'],
-    },
+    ...commonDirectiveOptions('math'),
   },
   body: {
     type: String,
     required: true,
   },
   run(data: DirectiveData): GenericNode[] {
-    const { label, identifier } = normalizeLabel(data.options?.label as string | undefined) || {};
-    const math = {
-      type: 'math',
-      identifier,
-      label,
-      value: data.body as string,
-    } as GenericNode;
+    const math = addCommonDirectiveOptions(data, { type: 'math', value: data.body as string });
     if (data.node.tight) {
       // The default `false` is not written to the AST
       math.tight = data.node.tight;